Let the assignment operator of abstract classes use the same convention as in non-abstract classes: - Accept &src == this. - Return *this instead of void.
In move constructors and move assignment operators, prefer exchange()
with an empty instance to std::move() + clear().

Implement assignment operator without explicit checks that &src == this.
Most of the time this just works because of how exchange() is used.
Sometimes this requires code reordering.

In the implementation of exchange(), do not make a local variable const.
This required the type to be copyable (even if NRVO elides the copy),
now it merely needs to be movable.

PiperOrigin-RevId: 182775160