Remove {,Backward}Writer::Fail().
They were a bit confusing: they had a slightly different semantics than
Object::Fail(). That semantics cannot be made thread-safe so they would
diverge even more (in fact I forgot about them during the change to
Object::Fail()).

Also, Fail() is not virtual, so they were shadowing Object::Fail()
rather than overriding it. This was a conscious decision and this worked
in practice, but this might add to confusion.

In effect particular writers ensure that start_ == cursor_ == limit_
on failure.

PiperOrigin-RevId: 182775784