If a transposed chunk would decode to a larger size than announced, abord decoding as soon as this can be detected instead of after decoding the whole chunk.
This will allow to protect against malicious input causing the decoder
to run out of memory by checking whether the announced size is small
enough.

(This is likely not enough, because apart from the concatenated values,
other parts of the decoded output might decompress to huge amounts of
data, e.g. the state machine. It is not clear how to derive the proper
limit of the state machine though.)

LimitingBackwardWriter was needed for this. LimitingWriter is provided
for completeness.

PiperOrigin-RevId: 187188058