In {Simple,Transpose}Decoder::Reset(), return sorted record end positions instead of boundaries, i.e. excluding the first 0.
This makes the convention consistent with ChunkEncoder::AddRecords() and
makes the number of limits equal to the number of records, removing
adding or subtracting 1 in various places.

The original motivation for passing both boundaries was to allow
computing record size in ChunkDecoder::ReadRecord() without a
conditional. But record start position can be obtained without a
conditional by Reader::pos().

Also, get rid of storing ChunkDecoder::num_records_. This is
limits_.size() now.

Also, fail if num_records and decoded_data_size are too large for the
given machine.

PiperOrigin-RevId: 188013637