Extract compressor options from RecordWriter::Options to a separate CompressorOptions class.
Use CompressorOptions instead of a fixed sequence of parameters
in Compressor and {Simple,Transpose}Encoder. This allows to avoid
specifying the defaults explicitly, and will make easier to add
other compression options without changing function signatures.

Let BrotliWriter::Options::kDefaultWindowLog() return
BROTLI_DEFAULT_WINDOW (22) instead of -1. This looks better in the
context of BrotliWriter (the only reason the default was mapped from -1
was that ZstdWriter has a different default and some classes specify
window_log which is then applied to either compressor), but requires
mapping CompressorOptions::kDefaultWindowLog() to different values
depending on the compressor.

In Compressor, extract computation of {Chain,Brotli,Zstd}Writer::Options
to separate functions. They are shared between Compressor constructor
and Reset().

Adding CompressorOptions required support from options parser: add
CopyOption() so that options can be split and some of them will be
parsed by another options parser.

Also, change OptionParser to return just the description of valid values
on error, with ParseOptions() building a full message from that. And add
AltOption(), so that parsing either a fixed value or a number can be
specified by composing EnumOption() with IntOption(). This fixes the
description of valid values for options brotli, zstd, and window_log.

PiperOrigin-RevId: 188729936