Add {Record,Chunk}Reader::corrupted_bytes_skipped() for reporting statistics about skipping corruption.
Fix reaction to corruption. When a chunk header has a valid hash but
implies an invalid chunk boundary, ensure that this is treated as a
failure or is skipped. Previously it might have been wrongly skipped
if recovery is used for seeking with skip_corruption_ being false.

Change the implementation of the reaction to use an invalid chunk
boundary. This is now detected before an attempt to read the chunk
header, so that it can be properly attributed to file corruption or not.

Fix reaction to truncation with HopeForMore() being true. Reset
is_truncated_ to false not after changing pos_ but at the beginning
of public functions. It should be reset even if pos_ is not changed,
e.g. after a failed Read() if HopeForMore() is false.

Change reaction to truncation with HopeForMore() being false. The
position is left at the end of file, instead of at a potential block
header after it, so that corrupted_bytes_skipped() does not report a
range which does not exist in the file.

When seeking to a block boundary which coincides with the end of the
file, do not bother verifying that a chunk ends there. There is no
obligation to verify contents of the file before the seek target.

Change the implementation of the beginning of recovery. Check if
corruption should be fatal before Recover(), not inside it, and update
the position to the next boundary unconditionally, even if this was
already done. This allows to remove the recovering_.in_progress
variable, which was somewhat confusingly used also to force recovery
even if corruption should be fatal. Introduce recovering_.corrupted
instead, to decide whether to update corrupted_bytes_skipped().

In RecordReader::Done(), do not destroy the ChunkReader, only Close()
it, so that corrupted_bytes_skipped() can be retrieved after closing
the RecordReader. Closing it might be the only moment where
corrupted_bytes_skipped() is updated if the file was truncated with
HopeForMore() being true but the file did not grow after all.

Add operator<<(ostream&, RecordPosition).

PiperOrigin-RevId: 189582265