Add RecordWriter::Pos() and an optional RecordPosition* parameter to RecordWriter::WriteRecord().
Rename RecordReader::pos() to Pos().

In parallel writing, leave the current position in the queue while it is
being handled. This is needed to compute the writing position. This has
a side effect that parallelism allows for one fewer threads than before;
set_parallelism(1) now encodes and writes only one chunk at a time, but
does so in background after WriteRecord() returns.

Prefer condition_variable::notify_all() over notify_one(). This is a
safer default, and they are equivalent if there is at most one thread
waiting (notify_one() should be treated as an optimization if there can
be multiple threads waiting but we know that only one will be able to do
useful work after unblocking).

PiperOrigin-RevId: 190620797