Rename {Record,Chunk}Reader::Options::set_skip_corruption() to set_skip_errors(), and {Record,Chunk}Reader::corrupted_bytes_skipped() to bytes_skipped(), because they will include not only corruption but also unparsable records, which do not necessarily indicate that the file is corrupted but that it is merely incompatible with expected record type.
Account for unparsable records (this applies to proto messages only)
in RecordReader::bytes_skipped().

Fix RecordReader::ReadRecordSlow(MessageLite*): this is called after a
failed ChunkDecoder::ReadRecord(), so ChunkDecoder might be unhealthy on
entry after all. OTOH in ReadRecordSlow() variants which do not parse a
proto message, assert that ChunkDecoder is healthy after a failed
ReadRecord(), not only on entry (if this assumption changes in future,
skipping errors would also need to be handled).

Remove the key parameter from ChunkDecoder::ReadRecord(). This can be
computed as ChunkDecoder::record_index() - 1 after calling ReadRecord().
This removes computation from the common path where key is nullptr.

PiperOrigin-RevId: 190771093