Move the parts of OptionsParser concerning parsing a single value to a separate ValueParser class.
Rename the old ValueParser type alias to ValueParser::Function.
Its parameter is now ValueParser* instead of string_view.

Add ValueParser::And().

Change ValueParser::FailIfSeen() and ValueParser::FailIfAnySeen() from
methods to be called from parsing functions to standalone parsing
functions, to be combined with other parsers with ValueParser::And().

Change the parameter of ValueParser::Empty() from std::function<bool()>
to T* out with T value.

This reduces the need of writing parsers as explicit lambdas. The only
remaining cases are in the first stage of CompressorOptions::Parse().

PiperOrigin-RevId: 192169989