Short Chain Optimization: represent Chains up to 16 characters (on 64-bit architectures) or 8 characters (on 32-bit architectures) without dynamic allocation.
This required changing the contract: moving a Chain may invalidate
its character pointers. ChainReader was relying on that.

Change the type of Chain::BlockIterator::reference to be the same as
value_type, with pointer being a proxy for operator->(), because the
pseudo-block pointing to short data does not contain a real string_view
to return a reference to.

Change the interface of Chain::BlockIterator::Pin() because it is no
longer possible to pin an arbitrary block while keeping its data pointer
unchanged. (Pinning small blocks is allowed but is rarely useful, it is
better to copy them by value, hence I am not worried that Pin() in such
case makes an allocation.)

Internal loops over blocks treat small data specially instead of using
BlockIterator, to hoist a conditional from the loop.

PiperOrigin-RevId: 192269233