When appending string_view to Chain, do not force it to be copied to a single block. Split it almost according to Chain block sizing rules, but do not start with very small blocks if the string_view is longer.
This is almost as fast given that block sizes are still not too small,
and it reduces memory pinned by potential subchains of the result.

Reduce maximum block capacity to numeric_limits<ptrdiff_t>::max().
Larger blocks are problematic, and even std::span cannot express them
(although we are using absl::Span for now which has an unsigned size).

PiperOrigin-RevId: 192598314