Do not set size_hint in Reader::Read(Chain*). This makes the common case up to 10% slower (usually 0%), but the worst case becomes much faster because in the worst case every operation would allocate.
Add a size_hint parameter to Append{,Partial}ToChain(MessageLite)
to compensate for this.

Add a missing size_hint argument inside
Chain::BlockIterator::AppendSubstrTo(Chain*).

Ideally we would have two notions of size_hint, the estimated final size
and the estimated minimal final size, but the interface complexity is
probably not worth this.

Also, ideally we would calculate the minimal size hint based on not what
will be the final size but what will be the size before the nearest
block which will be attached by reference.

PiperOrigin-RevId: 193232702