Move chunk type from the first byte of chunk data to a field in chunk header.
This will be useful to be able to look at a chunk header to decide
whether to read the chunk, e.g. for reading file metadata (to be
defined).

Ignore unknown chunk types if the chunk contains no records. This will
allow to add informational chunk types without breaking old decoders.

Split the notions of the file signature chunk and the padding chunk.
While they are treated the same way, they have a different purpose.

This is probably the last chance to change the file format in an
incompatible way.

PiperOrigin-RevId: 193658266