Change reporting invalid file contents or unparsable records in RecordReader, ChunkReader, and ChunkDecoder. Instead of having Options::set_skip_errors(true), they provide a Recover() function to recover from such failures, distinguishing them from other failures.
This will be necessary to be able to look at a chunk header to decide
whether to read the chunk:
- Reading file metadata (to be defined).
- Checking chunk size during parallel decoding (to be implemented).
Otherwise reading the chunk might have skipped errors and read a
different chunk than had its header checked. Also, this makes the
implementation simpler.

Recover() reports the number of skipped bytes or messages, instead of
having persistent counters. The number of skipped messages is reported
separately instead of counting a message as a number of bytes
corresponding to a difference of positions.

Remove checking the file format by RecordReader constructor.
Add RecordReader::CheckFileFormat().

Include more details in messages about invalid file contents.

Loosen an invariant of ChunkReader: byte_reader_->pos() can be less than
pos_ if a chunk boundary is known but we did not seek to it yet. Reading
a chunk seeks to it if needed. This simplifies code by consolidating
seeks, and makes the solution from the next paragraph possible.

In ChunkReader::Seek() to a position past the end of file, report
success, with pos() returning the requested position, instead of
reporting failure and assuming that the physical end of file is a chunk
boundary. This improves the behavior for a growing file ending with a
truncated chunk: after the file grows, the end of file is no longer a
chunk boundary, so growing could have invalidated the position. Also,
since it is now possible that pos() returns a position past the end of
file, it should be possible to seek back to it.

ChunkReader::Seek() now fails when the position is an invalid chunk
boundary. Accepting such positions (with a special behavior during
reading) was needed to be able to seek again to pos() returned after
seeking past the end of file, but this is no longer needed.

Remove the special behavior of ChunkReader when the current position,
which is normally a chunk boundary, is a block boundary or an invalid
chunk boundary position. The special behavior was to find the first
chunk in this block and skip there. Motivation was: when recovering in
a growing file which does not have a valid chunk beginning yet, pos()
should return a position which continues the recovery; also after
seeking past the end of file, the current position is not necessarily
a valid chunk boundary. Recover() may now still leave the ChunkReader
at a block boundary, but ReadChunk() now verifies that a chunk begins
there, and if not, reports invalid file contents again instead of
skipping to a chunk. This preserves the property from the motivation.

In ChunkReader::SeekToChunk{Containing,After}(), start with the last
full block header when seeking past the end of file, or when seeking to
a block with an incomplete block header at the end of file. This ensures
that it finds a chunk boundary even if the file is truncated (but not
when the file is corrupted).

PiperOrigin-RevId: 195073445