Remove comments about constructors of inner classes not being defaulted. Do not mark Options constructors as constexpr.
No constructors are defaulted now for another reason (errors are
reported too late if defaulting happens at the declaration) so this is
no longer a special case.

constexpr is a promise which binds future implementations (i.e. removing
it is a potentially incompatible change), while these constructors may
happen to be not constexpr in future, and constexpr is not that useful
for them.

PiperOrigin-RevId: 195954085