Add optional file metadata to Riegeli/records files, mainly to inform about the type of records, expressed as a RecordsMetadata proto message.
File metadata is written by RecordWriter::Options::set_metadata() and
read by RecordReader::ReadMetadata().

Add SetRecordType() function for setting metadata from a message
descriptor of the type of records, and RecordsMetadataDescriptors
class for getting a message descriptor from metadata.

Allow for non-zero decoded_data_size when num_records is 0. This is
useful for the metadata chunk.

Move parsing of chunks with no records from RecordReader back to
ChunkDecoder. There was a plan for RecordReader to retain information
from them, which was the rationale of the previous arrangement, but it
changed.

PiperOrigin-RevId: 196497731