Let writers to string-like objects optionally own their destination. This is requested by passing kOwnsDest() instead of a destination pointer (to keep the convention that the default constructor creates a closed Writer). This is not merely a convenience; this allows them to be movable together with their destination, which would otherwise be tricky.
Everything applicable to Writers applies also to BackwardWriters.

For completeness, expose src() and dest() of various Readers and
Writers, no matter whether they are owned. Also, expose fd() of
FdReaders and FdWriters. Since dest() of a Writer which owns its
destination must be available and unchanged after Close(), let all these
accessors work after Close(), except in cases where that would be
meaningless (closing a fd).

Remove the precondition that when a FdReader or FdWriter opens the file
itself, it must be owned. While not owning it in this case is unusual,
it is no longer necessarily a resource leak (the file is still
accessible by fd()).

Let FdMMapReader keep the fd open, so that fd() works consistently with
FdReader. Add FdMMapReader::Options::set_sync_pos().

Close() is going to leave more of the object state unchanged:

* Let pos() and RecordWriter::Pos() be unchanged by Close(). Subtracting
  positions will replace getting the size of the corrupted region by an
  output parameter of Recover(), and Recover() is meaningful after
  Close().

* Do not bother with cleaning all object state in Done() overrides
  (there were already exceptions). Owned subobjects are closed but not
  destroyed; sometimes this is needed e.g. for pos().

* Remove a comment that Close() frees all owned resources since this is
  no longer true.

* Do not force Done() to be implemented by all subclasses (it was
  abstract in Object, and implemented but still abstract in abstract
  classes like Reader and BufferedReader), since it is increasingly
  valid to leave a derived Done() doing nothing.

When stating that closing an Object closes resources it owns, do not
make this conditional only when the Object is healthy. This used to be
applied inconsistently, and it seems simpler to do it unconditionally.
In practice this does not matter much, because client code is unlikely
to close objects after failure; if this is done, it should not hurt that
e.g. file buffers containing previously prepred chunks are written to
the file if serializing a record in the next chunk failed.

Clean up comments. In particular mark accessors which are valid after
Close().

Remove the invariant that a closed ChainReader has an iterator pointing
to an empty Chain, because owned_src_ is unchanged by Close() and thus
ChainReader no longer has a handy empty Chain to point the iterator to.
This requires having explicit checks for healthy(), as in most other
Readers and Writers, which could have been spared before because the
following code happened to do nothing in case it was reading from an
empty Chain.

In ChainReader move constructor and move assignment, do not bother with
computing cursor index (i.e. cursor_ - start_) before moving the Chain.
Assume that even after a Chain is moved, pointers which previously
belonged to it can still be subtracted (they are not dereferenced).
This is now also the case for ChainWriter.

Simplify RecordWriter initialization: ensure that whenever
RecordWriter::Impl is created, its chunk_encoder_ != null
(which is needed for Pos()).
PiperOrigin-RevId: 198371646