Do not bother accepting the filename to open as string by value. Accept it as string_view.
This simplifies calling code which has the filename as a string_view.
While this adds a memory allocation and copy if the caller has a string
rvalue, this is unlikely to matter in practice because opening a file is
much more expensive than that, and filenames tend to be short.

In theory we might have 3 overloads (string_view, string&&, const char*)
but the added API complexity is not worth the optimization.

PiperOrigin-RevId: 211854015