Inline functions more aggressively.
In particular mark template functions inline if they would be defined
inline were they not templates. Especially move constructors and move
assignment operators, and private Move{Src,Dest}() functions.

While adding 'inline' at 'template' does not change the semantics
(the function already can be defined in multiple translation units),
it does influence compiler inlining heuristics in both clang and gcc:
https://blog.tartanllama.xyz/inline-hints/

This is especially true for members of class templates for which some
instantiations are declared as 'extern template class' and defined as
'template class'. Without 'inline' clang seems to never inline them,
even though it does inline a member of an instantiation which is not
declared extern in the header. Declaring them extern was meant to reduce
linker work in case the compiler decides to not inline a member, but not
to prevent inlining of all members.

With 'inline' the compiler still uses heuristics, and avoids inlining
the function if its compiled code gets too big.

Details:

- constexpr formally implies inline, so do not bother marking constexpr
  functions inline. Mark Saturating{Add,Sub}() constexpr instead of
  inline.

- {Brotli,Zstd,Zlib}{Reader,Writer} constructors from {src,dest}, except
  for ZstdWriter, unchanged (nor marked inline) because they are
  expensive anyway.

- ChainReaderBase::Initialize() manually inlined into ChainReader
  constructor.

- Fd{,Stream,MMap}{Reader,Writer} constructors from {src,dest} unchanged
  (not marked inline) because they are expensive anyway.

- Do not bother inlining
  Limiting{,Backward}Writer::Supports{RandomAccess,Truncate}().
  They are virtual.

- ZstdWriterBase::operator=() defined inline, even though it calls
  InitializeCStream() (which is no longer inline in zstd_writer.cc),
  because the call is conditional. Let the compiler decide.

- ChunkDecoder constructor and moving defined inline. Let the compiler
  decide.

PiperOrigin-RevId: 212979363