Change Limiting{Reader,Writer,BackwardWriter} to optionally own their source or destination, like other Readers and Writers.
Rename SetSizeLimit() to set_size_limit(). For completeness, add it to
Limiting{,Backward}Writer.

Make the size limit optional, defaulting to infinity. This is useful
when set_size_limit() is used later.

Remove the optimization where if the immediate source of a
LimitingReader is a LimitingReader too, it reads directly from the real
source. It was too painful to maintain. In fact the interaction between
this optimization and set_size_limit() was incorrect: it was possible to
read the real source past the size of the immediate source by setting a
higher limit.

Add SizeLimitSetter to set the size limit of a LimitingReader in the
constructor and restore it in the destructor. This is more efficient
than making a new LimitingReader reading from a LimitingReader.

When iterating through a Reader using LimitingReader, use a single
LimitingReader with a repeated calls to set_size_limit() instead of
constructing a separate LimitingReader for each iteration.

Fix class invariants in comments. They were incorrect when
set_size_limit() is used.

Fix typos in assertion messages.

PiperOrigin-RevId: 218656259