Allow FdStreamReader which does not own the fd.
Only warn that it will have an unpredictable amount of extra data
consumed because of buffering.

While this can be problematic, this is not unconditionally useless.
Even if the caller will not read from the file further, it might
still prefer to close it due to how other libraries are organized.

Python allows open(fd, mode, closefd=False), even though it has the same
issue with buffering.

PiperOrigin-RevId: 221229621