Use absl::aligned_storage_t<sizeof(T), alignof(T)> storage_; instead of union { T object_; }; for the storage in NoDestructor<T>.
This ensures that NoDestructor has a trivial destructor, not merely a
destructor which does nothing. It should make no difference in practice,
but it allows to meet the constraint that variables with static storage
duration have trivial destructors.

This forced to remove constrexpr support from NoDestructor. That support
was currently unused (all NoDestructor usages have local scope).

PiperOrigin-RevId: 222399180