Move the responsibility for ensuring that Fail() makes the buffer empty from individual derived classes to Writer and BackwardWriter.
This required making Fail() virtual.

This used to be impossible because some classes owned resources held
only by their buffer pointers, but this is no longer the case.

There is no corresponding guarantee for Reader because the buffer of a
Writer corresponds to future writes, while the buffer of a Reader
corresponds to past reads. If the destination of a Writer failed, there
nothing more can be written, but if the source of a Reader failed, what
has been already read can be seen.

PiperOrigin-RevId: 223935261