Abort if assumed_pos is not specified when it is required.
Constructors of non-seekable FdStream{Reader,Writer} from an already
open file do not know the initial position, and knowing the position is
essential for Riegeli/records format, so Options::set_assumed_pos() is
required to force the caller to consider the initial position.

But this is not obvious and easy to miss, so change RIEGELI_ASSERT to
RIEGELI_CHECK, to include the check also in opt mode.

PiperOrigin-RevId: 230943109