Use linkstatic = True to link the helper py_binary used by py_extension.
This fixes running Riegeli-dependent Python binaries when Riegeli is
built from an external repository (and the Python directory is not in a
root directory of a workspace), by working around a subtle problem with
finding dynamic libraries relatively to $ORIGIN when a shared object is
accessed via a symlink.

This causes linking the same code multiple times, but at least it works.

PiperOrigin-RevId: 231945829