Add SizeReturningNewAligned() which couples size rounding with memory allocation, to make it easier to port this to a memory allocator implementing https://wg21.link/P0901
Allow Buffer::GetData() to update Buffer::size() to account for size
rounding by the underlying memory allocator.

Remove the optimization that Buffer assignment reused the destination
allocation if the source was not allocated and the size matched. This is
unlikely to happen if allocation changes the size. Fortunately it was
not terribly important.

PiperOrigin-RevId: 235892300