Tune the block sizing strategy in Chain.
Increase the minimum block size from 128 to 256, and increase the growth
ratio from 1.5 to 2, i.e. a new block has the size of the Chain so far,
not half of that, but no more than 64 KB - this limit is unchanged.

This indirectly makes the criterion of wasteful blocks slightly more
permissive, because it involves the minimum block size.

PiperOrigin-RevId: 240423689