Update version requirements.
Require Python 2.7 or >=3.5, <4.

Require protobuf>=3.6.0. Generated *_pb2.py code is compatible with
protobuf 3.6.0 but not with 3.5.2.

Allow tensorflow>=1.14 instead of >=2.0.0a0. Riegeli should be
compatible with TensorFlow 1.14 when it is released.

Declare incompatibility with protobuf 4 and tensorflow 3, should they
ever be released.

Revert to 'distclass=BinaryDistribution' instead of
'ext_modules=FakeNonEmptyList()'. According to
https://www.python.org/dev/peps/pep-0491/#what-s-the-deal-with-purelib-vs-platlib
  A wheel with "Root-Is-Purelib: false" with all its files in
  {name}-{version}.data/purelib is equivalent to a wheel with
  "Root-Is-Purelib: true" with those same files in the root, and it is
  legal to have files in both the "purelib" and "platlib" categories.
so despite "purelib" appearing in paths, this should be fine.

PiperOrigin-RevId: 242080584