Use string::erase() instead of string::resize() for reducing the size.
This generates better code. While in general erase() can move the part
of the string after the removed part, the compiler can see that the
count parameter of erase() is npos (defaulted), so there is nothing to
move. There is only a conditional jump to throwing an exception, which
should never happen. With resize() the compiler cannot rule out that the
size grows.

PiperOrigin-RevId: 245385341