Remove private functions Chain::Block::PrepareFor{Append,Prepend}().
Instead, perform their task (namely shifting the block contents to the
appropriate end of the block if there is nothing to shift) in functions
which do the append/prepend or check how much can be appended/prepended.

This removes the need to check the block kind and the atomic reference
count twice.

PiperOrigin-RevId: 245404082