Let {Read,Copy}Varint{32,64}() utilize Pull() with min_length.
This eliminates the slow path of these functions, at the cost of making
the fast path a bit slower: the limit must be checked at each byte.
Also, this pulls more data than might be necessary for parsing the
varint.

This is a tradeoff, not obviously a win like it was in
WriteVarint{32,64}(). It turns out that this is faster for short varints
and slower for long varints.

Other changes:

- In ReadVarint{32,64}() variants which read from an array, add a limit
  parameter instead of assuming that at least kMaxLengthVarint{32,64}
  bytes are available.

- Add CopyVarint{32,64}() variants which read from an array. This makes
  CopyVarint{32,64}() analogous to ReadVarint{32,64}().

- In TransposeDecoder, when processing a string field, read its length
  using Pull() with min_length.

PiperOrigin-RevId: 248296143