Support field projection in all cases.
Field projection is currently broken in extremely rare cases:
- Nesting more than 100 levels deep.
- Non-shortest encoding of a field tag, varint value, or string length.

In these cases the projection could have skipped fields which were
supposed to be included. There was no problem with deep nesting inside
a projected field, the problem was only when projection specified
deeply nested fields.

After this change the whole submessage (either deeply nested or
suboptimally encoded) will be included if any of its fields are
specified in the projection. This satisfies the projection contract
since projection is best effort.

PiperOrigin-RevId: 248711559