Improve performance of Chain::{Apppend,Prepend}Buffer() in some cases.
If the caller asks for a large buffer but fills it only partially, and
does that multiple times, the same block might be wasteful again, and
the same data could be rewritten multiple times (until the block gets
large enough relatively to its wasted space that the block is no longer
considered wasteful).

Solution: rewrite the wasteful block separately from the new space. This
needs one more allocation, but the data gets rewritten to a non-wasteful
block only once.
PiperOrigin-RevId: 249424897