Change function templates with a parameter type convertible to string_view back to plain functions with a const char* parameter. This is simpler to understand.
It is likely that const char* is the only type convertible to both
string_view and string which may realistically appear here. If another
such type emerges, the ambiguity can be resolved explicitly by the
caller.

It is not even obvious that resolving to string_view is better if a type
is destructively convertible to string in an efficient way.

PiperOrigin-RevId: 250110143