Avoid unnecessary moves.
Objects with dependencies owned by value used to be constructed like
this:

  BrotliWriter<ChainWriter<Chain>> writer(
      (ChainWriter<Chain>(Chain()));

with extra parens sometimes needed due to the most vexing parse.

The effect forced by interfaces was:

  - Construct a Chain.
  - Construct a ChainWriter<Chain>:
    - Move the Chain to the ChainWriter.
  - Construct a BrotliWriter<ChainWriter<Chain>>:
    - Move the ChainWriter<Chain> to the BrotliWriter.
      - Move the Chain in ChainWriter, updating buffer pointers.

The actual effect was even worse, because of absl::exchange() used in
DependencyBase: moving a ChainWriter<Chain> moved the Chain twice,
moving a BrotliWriter<ChainWriter<Chain>> moved the ChainWriter twice
and thus the Chain 4 times, etc.

Unfortunately the moves are not as cheap as once thought. And they are
not really necessary: they are performed because the target object is
described by constructing a temporary object first. While this affects
only object construction, sometimes this matters, e.g. when
ParseFromReader() is used in a loop.

Hence there are 3 major changes:

1. Whenever a constructor takes a dependency to move from it, it can
   also take a tuple of parameters for constructing the dependency in
   place. Such tuple is usually written std::forward_as_tuple(args...).

   This mimics forwarding constructors of other types, like:

     std::optional::optional(std::in_place_t, args...)

     std::pair::pair(std::piecewise_construct_t,
                     std::tuple(args1...), std::tuple(args2...))

   except that the arguments are always wrapped in a tuple (so that
   other optional parameters can follow), and the constructor is not
   tagged to disambiguate it from other constructors (because
   dependencies have concete types which are not tuples themselves).

2. For completely replacing the state of an object, instead of assigning
   a newly constructed object to it, prefer object->Reset(args...),
   or another construct depending on the type, or
   riegeli::Resetter<T>(object, args...) in generic code.

   The Reset() method mimics std::string::assign() which is overloaded
   like std::string constructors. Such methods can be more efficient
   than constructing and moving from a temporary object, not only by
   avoiding the overhead of moving but also by reusing already allocated
   storage.

3. The dependency is taken by a const reference or rvalue reference,
   instead of by value.

   This eliminates an extra move if the dependency is std::move(lvalue).

These changes can be combined: Reset() can take a tuple, etc.

The preferred way to construct an object with dependencies owned by
value is now:

  BrotliWriter<ChainWriter<Chain>> writer(
      std::forward_as_tuple(std::forward_as_tuple());

The effect is:

  - Construct a BrotliWriter<ChainWriter<Chain>>:
    - Construct the ChainWriter<Chain> in the BrotliWriter.
      - Construct the Chain in the ChainWriter.

For functions rather than classes, like SerializeToWriter() and
ParseFromReader(), specify a template parameter. Instead of this:

  Status parse_status = ParseFromReader(
      record, LimitingReader<>(&values_reader_, limit));

it is better to write this:

  Status parse_status = ParseFromReader<LimitingReader<>>(
      record, std::forward_as_tuple(&values_reader_, limit));

And for delayed initialization after construction, instead of this:

  reader = ChainReader<>(&src);

it is better to write this:

  reader.Reset(&src);

Also, explicit instantiations of classes for common template arguments
are removed. They interfered with some members being inapplicable for
some template arguments, e.g. trying to copy a unique_ptr. While it
would have been possible to mark the problematic constructors and
methods with enable_if, it was simpler to remove explicit
instantiations.

PiperOrigin-RevId: 251885550