Add missing declarations for explicit instantiations.
The current behaviour (using externally defined instantiations without having seen a declaration of those external instantiations) is undesirable and effectively deprecated, and is warned about by -Wundefined-func-template. (We don't enable this flag by default yet because of the large number of extant violations, but would very much like to.)

Found with -Wundefined-func-template.

PiperOrigin-RevId: 252017511