Optimization in Chain: change the representation of special values of pointers in the block pointer array (kBeginShortData and kEndShortData).
Instead of using pointers in a static dummy array, use constants with
values 0 and sizeof(Block*). This allows to load their values without
referring to the address of a static object, and for a typical pointer
representation the special values differ by as much as real pointers.

Pointers and constants are internally represented by intptr_t, because
pointers can be cast to intptr_t and back, but conversely arbitrary
integers may not be cast to a pointer and back. This is wrapped in a
private BlockPtrPtr type.

This design forces treating the special values separately in the source
code, to avoid a non-portable assumption that pointer arithmetic
corresponds to scaled integer arithmetic. For a typical pointer
representations the correspondence indeed holds, the compiler can
realize that both branches have the same behavior, and eliminate the
checks.

PiperOrigin-RevId: 252292452