Let Dependency::is_owning() return false when the dependency does not exist, e.g. when it is held by unique_ptr which is nullptr.
Do not check whether an owned dependency exists in Done() of file
readers and writers.

The check in file readers and writers has been added so that after a
failed open we do not try to close a non-existent file handle. It is
more robust to have the check in is_owning().

PiperOrigin-RevId: 252602131