Do not clear the state of moved-from objects unless this is necessary to maintain invariants or to avoid catastrophic failures when such object is used in certain ways.
Usually it is sufficient that the moved-from object is closed if it
derives from Object, because most operations are guarded by healthy().

The moved-from state is still cleared when fields are managing ownership
(e.g. are accessed in an explicitly written destructor), and the
moved-from state ensures that Recover() does not try to recover
(Recover() is applicable even to a closed object).

In ChunkDecoder it used to be important that a moved-from object is
empty, because RecordReaderBase::ReadRecord() decodes a record before
any health checks, and ChunkDecoder used to not derive from Object. But
this is no longer needed: ChunkDecoder checks its own healthy() before
trying to decode a record.

PiperOrigin-RevId: 252662392